% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre.R
\name{data_pre}
\alias{data_pre}
\title{Data_pre: pre-process data for regression}
\usage{
data_pre(ln_name, fc_name, data)
}
\arguments{
\item{ln_name}{Variable you want to ln}

\item{fc_name}{Variable you want to treat as factor}

\item{data}{Data waitted ti be compute}
}
\value{
A transformed dataframe with factored out variables and log improvement
}
\description{
This function is used to pre-process the data for regression.
For variables that require categorical analysis, this function factors out the variable.
If the data distribution is too skewed, log improvement is used.
This function will return the changed data frame
}
\examples{
data("test")
ln_name <- c("AGE")
fc_name <- c("ASTHMA")
data <- test
data_pre(ln_name,fc_name,data)

}
\references{
Henry L, Wickham H (2020). purrr: Functional Programming Tools. R package version 0.3.4, \url{https://CRAN.R-project.org/package=purrr}.

Wickham H, François R, Henry L, Müller K (2022). dplyr: A Grammar of Data Manipulation_. R package version 1.0.10, \url{https://CRAN.R-project.org/package=dplyr}.
}
