% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{Describe}
\alias{Describe}
\alias{describe}
\title{Describe: summarize the basic information of data}
\usage{
describe(x, y, cov, data)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Dependentvariable}

\item{cov}{Covariates}

\item{data}{Dataset you want to describe,must be a dataframe}
}
\value{
A table in Viewwe windows,and summary statistics of variables
}
\description{
This function is used to describe
the basic situation of the variables in
the data set, such as the maximum and minimum values,
the 25th quartile and the 75th quartile.
A summary list is also returned in the result,
and this table shows the detailed distribution of the
variables entered. Variables need to be specified as
independent, dependent and covariate variables
The data set contains the prevalence of asthma and some
basic information about the population such as
BMI, activity level (PA), age, gender, race, alcohol use (ALQ), and smoking status (COT).
}
\examples{
data("test")
y <- c("ASTHMA")
x <- c("AGE")
cov <- c("PA","BMI")
data <- test
describe(x,y,cov,data)
}
