% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_process.R
\name{glm_process}
\alias{glm_process}
\title{Glm_press}
\usage{
glm_process(x, y, cov, family, data)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Dependentvariable}

\item{cov}{Covariates}

\item{family}{Binomial or gaussian}

\item{data}{Data to be computed}
}
\value{
A result dataframe and regression curves with points of x,use plot() to draw plots
}
\description{
This function is used to fit the linear relationship between variables
generalize linear model  an extension of the linear model that
mathematical expectations of the response variables and the predictor variables of the linear combination are established by a linkage function
The relationship between the mathematical expectation of the response variable and the linear combination of the predictor variables.
It is an extension of the linear model in studying the non-normal
distribution of response values and the concise and direct linear transformation of nonlinear models.
}
\examples{
data("test")
y <- c("ASTHMA")
x <- c("AGE")
cov <- c("PA","BMI")
family <- "binomial"
data <- test
glm_process(x,y,cov,family,data)




}
