% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runasthma.R
\name{runasthma}
\alias{runasthma}
\title{Launch Shiny App for asthma}
\usage{
runasthma()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for asthma
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
asthma::runasthma()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
