% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Cor_plot: fits the Spearman's correlation between variables}
\usage{
cor_plot(cor_name, data)
}
\arguments{
\item{cor_name}{Variables you want to make corplot}

\item{data}{Data waitted ti be compute}
}
\value{
returns a corrplot demonstrates the Spearman's correlation relationship between variables.
}
\description{
This function is used to fit the correlation between variables
The Spearman's correlation coefficient indicates the direction of correlation between X (the independent variable) and Y (the dependent variable). If
If Y tends to increase when X increases, the Spearman's correlation coefficient is positive. If Y tends to decrease when X
increases, Y tends to decrease, the Spearman correlation coefficient is negative. Spearman's correlation
A zero coefficient indicates that there is no tendency for Y to increase when X increases. As X and Y get closer to complete
The Spearman correlation coefficient increases in absolute value as X and Y become closer to complete monotonic correlation. When X and Y
are completely monotonically correlated, the absolute value of the Spearman correlation coefficient is 1.
This function returns a plot of the total correlation, with the size and color of the circles showing the correlation
}
\examples{
data("test")
data <- test
cor_name <- c("PA","BMI","AGE")
cor_plot(cor_name,data)

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Taiyun Wei and Viliam Simko (2021). R package 'corrplot': Visualization of a Correlation Matrix (Version 0.92). Available from \url{https://github.com/taiyun/corrplot}

Kassambara A (2022). ggcorrplot: Visualization of a Correlation Matrix using 'ggplot2'. R package version 0.1.4, \url{https://CRAN.R-project.org/package=ggcorrplot}.

R Core Team (2022). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. \url{https://www.R-project.org/}.
}
