% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre.R
\name{data_pre}
\alias{data_pre}
\title{Data_pre}
\usage{
data_pre(ln_name, fc_name, data)
}
\arguments{
\item{ln_name}{Variable you want to ln}

\item{fc_name}{Variable you want to treat as factor}

\item{data}{Data waitted ti be compute}
}
\value{
A transformed dataframe
}
\description{
This function is used to pre-process the data for regression.
For variables that require categorical analysis, this function factors out the variable.
If the data distribution is too skewed, log improvement is used.
This function will return the changed data frame
}
\examples{
data("test")
ln_name <- c("AGE")
fc_name <- c("ASTHMA")
data <- test
data_pre(ln_name,fc_name,data)

}
