% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his_plot.R
\name{his_plot}
\alias{his_plot}
\title{his__plot: demonstrates the distribution of chosen variable across population.}
\usage{
his_plot(x, title, xlab, ylab, data)
}
\arguments{
\item{x}{Character vector specifying the variable from the dataset to be plot. Example "AGE".}

\item{title}{Plot's title}

\item{xlab}{Name of xlab}

\item{ylab}{Name of ylab}

\item{data}{to be compute}
}
\value{
A histogram plot with variable x and its relative frequency of occurrance in the samples.
}
\description{
This function serves to draw a histogram of the variables
A histogram is an exact graphical representation of the distribution of numerical data.
It is an estimate of the probability distribution of a continuous variable (quantitative variable) that
The first step is to segment the range of values
These values are usually specified as continuous, non-overlapping intervals of the variable.
The intervals must be adjacent and usually (but not necessarily) of equal size
}
\examples{
data("test")

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
